aucat: Use extended header format for .wav files.
According to Microsoft docs, it is needed if bits > 16 or if there are
more than 2 channels, which aucat supports and is the defaut. Fixes
errors reported by sox when trying to play .wav files generated by
aucat.

Reported by John Rigg <obsd at jrigg.co.uk> and others on misc@